Mohan - Payment Service - Initial Implementation of Payment Service

+++ src/main/java/com/sstpath/marketplace/payment/dto/request/ChipPurchaseRequest.java
@@ -1,8 +1,9 @@
 package com.sstpath.marketplace.payment.dto.request;
 
-import lombok.Data;
-
+import java.math.BigDecimal;
 import java.util.List;
+
+import lombok.Data;
 
 @Data
 public class ChipPurchaseRequest {
@@ -25,7 +26,7 @@ public class ChipPurchaseRequest {
     @Data
     public static class PurchaseDetails {
         private List<Product> products;
-        private int total;
+        private BigDecimal total;
         private String currency;
         private String invoice;
